test: add test coverage to fs.promises.utimes and filehandle.utimes
